backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.8.3
aiosignal==1.3.1
altair==5.5.0
altgraph==0.17.4
amqp==5.3.1
asgiref==3.8.1
asn1crypto==0.24.0
async-timeout==4.0.3
attrs==24.2.0
Babel==2.6.0
beautifulsoup4==4.13.5
billiard==4.2.1
blinker==1.9.0
boto3==1.35.76
botocore==1.35.76
cachetools==5.5.1
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
channels==4.1.0
chardet==3.0.4
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
configparser==3.7.3
contourpy==1.3.1
cron-descriptor==1.4.5
cryptography==45.0.6
curl_cffi==0.12.0
cycler==0.12.1
deep-translator==1.11.4
distlib==0.3.9
dj-database-url==2.3.0
Django==5.2.6
django-admin-interface==0.30.1
django-adminlte-3==0.1.6
django-adminlte3==0.1.6
django-celery-beat==2.7.0
django-colorfield==0.14.0
django-cors-headers==4.8.0
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-filter==24.3
django-haystack==2.8.1
django-jazzmin==3.0.1
django-jet-reboot==1.3.10
django-phonenumber-field==2.2.0
django-redis==5.4.0
django-storages==1.14.4
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
dlib==19.24.6
dnspython==2.7.0
enum34==1.1.6
et_xmlfile==1.0.1
filelock==3.18.0
fonttools==4.56.0
frozendict==2.4.6
frozenlist==1.4.1
fsspec==2025.3.2
future==0.17.1
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
huggingface-hub==0.30.2
idna==2.8
jdcal==1.4
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
kombu==5.5.4
langdetect==1.0.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdocument==4.0.1612576120
mdurl==0.1.2
mlxtend==0.23.4
motor==3.7.1
mpmath==1.3.0
multidict==6.1.0
multitasking==0.0.12
mysqlclient==2.2.6
narwhals==1.25.2
networkx==3.4.2
numpy==2.2.4
oauthlib==2.1.0
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
openpyxl==2.5.12
packaging==25.0
pandas==2.2.3
pdf2image==1.17.0
peewee==3.18.2
pefile==2023.2.7
phonenumbers==8.10.6
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.2.0
protobuf==3.6.0
psycopg==3.2.12
psycopg-binary==3.2.12
psycopg2-binary==2.9.10
pur==7.3.3
pyarrow==19.0.0
pycparser==2.19
pydeck==0.9.1
Pygments==2.19.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.0
PyJWT==1.7.1
pymongo==4.13.2
PyMySQL==0.9.3
pyOpenSSL==19.0.0
pyparsing==3.2.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.16.1
PySocks==1.6.8
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-slugify==8.0.4
pytz==2018.7
pywin32-ctypes==0.2.3
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
requests-oauthlib==1.0.0
rich==13.9.4
rpds-py==0.22.3
s3transfer==0.10.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
soupsieve==2.8
sqlparse==0.5.3
streamlit==1.42.0
sympy==1.13.1
tenacity==9.0.0
text-unidecode==1.3
threadpoolctl==3.5.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.51.1
twilio==9.3.4
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.32.0
vine==5.1.0
virtualenv==20.28.0
watchdog==6.0.0
wcwidth==0.2.13
websockets==15.0.1
whitenoise==6.7.0
Whoosh==2.7.4
xlrd==1.2.0
yarl==1.15.4
yfinance==0.2.65
